requirements.txt
django
django-rest-framework
pillow
rembg[cpu]